fix window height & rename Wrapper for input elements
